requirements.txt
argon2-cffi==19.2.0
faker==4.0.1
Flask==1.1.1
flask-jwt-extended==3.24.1
passlib==1.7.2
pytz==2019.3
psycopg2==2.8.4
requests==2.23.0
structlogger==20.1.0

ipython==7.13.0
pdbpp==0.10.2
colorama==0.4.3
